# client: fix the pd client could be blocked in some cases