--- a/client/client.go
+++ b/client/client.go
@@ -140,8 +140,8 @@ func WithRetry(retry uint64) RegionsOption {
 
 type tsoRequest struct {
 	start      time.Time
-	client     *client
-	ctx        context.Context
+	clientCtx  context.Context
+	requestCtx context.Context
 	done       chan error
 	physical   int64
 	logical    int64
@@ -428,7 +428,7 @@ func (c *client) createTSODispatcher(dcLocation string) {
 
 func extractSpanReference(requests []*tsoRequest, opts []opentracing.StartSpanOption) []opentracing.StartSpanOption {
 	for _, req := range requests {
-		if span := opentracing.SpanFromContext(req.ctx); span != nil {
+		if span := opentracing.SpanFromContext(req.requestCtx); span != nil {
 			opts = append(opts, opentracing.ChildOf(span.Context()))
 		}
 	}
@@ -507,7 +507,7 @@ func tsLessEqual(physical, logical, thatPhysical, thatLogical int64) bool {
 
 func (c *client) finishTSORequest(requests []*tsoRequest, physical, firstLogical int64, err error) {
 	for i := 0; i < len(requests); i++ {
-		if span := opentracing.SpanFromContext(requests[i].ctx); span != nil {
+		if span := opentracing.SpanFromContext(requests[i].requestCtx); span != nil {
 			span.Finish()
 		}
 		requests[i].physical, requests[i].logical = physical, firstLogical+int64(i)
@@ -567,8 +567,8 @@ func (c *client) GetLocalTSAsync(ctx context.Context, dcLocation string) TSFutur
 		ctx = opentracing.ContextWithSpan(ctx, span)
 	}
 	req := tsoReqPool.Get().(*tsoRequest)
-	req.ctx = ctx
-	req.client = c
+	req.requestCtx = ctx
+	req.clientCtx = c.ctx
 	req.start = time.Now()
 	req.dcLocation = dcLocation
 	c.waitForDispatcher()
@@ -630,10 +630,10 @@ func (req *tsoRequest) Wait() (physical int64, logical int64, err error) {
 		cmdDurationWait.Observe(now.Sub(start).Seconds())
 		cmdDurationTSO.Observe(now.Sub(req.start).Seconds())
 		return
-	case <-req.ctx.Done():
-		return 0, 0, errors.WithStack(req.ctx.Err())
-	case <-req.client.ctx.Done():
-		return 0, 0, errors.WithStack(req.client.ctx.Err())
+	case <-req.requestCtx.Done():
+		return 0, 0, errors.WithStack(req.requestCtx.Err())
+	case <-req.clientCtx.Done():
+		return 0, 0, errors.WithStack(req.clientCtx.Err())
 	}
 }