--- a/client/client_test.go
+++ b/client/client_test.go
@@ -117,26 +117,26 @@ type testTsoRequestSuite struct{}
 
 func (s *testTsoRequestSuite) TestTsoRequestWait(c *C) {
 	ctx, cancel := context.WithCancel(context.Background())
-	cli := &client{
-		baseClient: &baseClient{
-			urls:            []string{"localhost:8080"},
-			checkLeaderCh:   make(chan struct{}, 1),
-			ctx:             ctx,
-			cancel:          cancel,
-			security:        SecurityOption{},
-			gRPCDialOptions: []grpc.DialOption{grpc.WithBlock()},
-		},
-		checkTSDeadlineCh: make(chan struct{}),
-	}
-
-	tsoRequest := &tsoRequest{
-		done:     make(chan error, 1),
-		physical: 0,
-		logical:  0,
-		ctx:      context.TODO(),
-		client:   cli,
+	req := &tsoRequest{
+		done:       make(chan error, 1),
+		physical:   0,
+		logical:    0,
+		requestCtx: context.TODO(),
+		clientCtx:  ctx,
 	}
 	cancel()
-	_, _, err := tsoRequest.Wait()
+	_, _, err := req.Wait()
+	c.Assert(errors.Cause(err), Equals, context.Canceled)
+
+	ctx, cancel = context.WithCancel(context.Background())
+	req = &tsoRequest{
+		done:       make(chan error, 1),
+		physical:   0,
+		logical:    0,
+		requestCtx: ctx,
+		clientCtx:  context.TODO(),
+	}
+	cancel()
+	_, _, err = req.Wait()
 	c.Assert(errors.Cause(err), Equals, context.Canceled)
 }