@leoppro 2020-12-17
> @leoppro Could you please bring this change to `release-4.0` also? The PD client in that branch should have the same problem.

ok