@Yisaer 2020-12-17
> There are two levels of context in PD client: client context and tso request context.

I think when `clientCtx` is closed, the caller should guranteed the `requestCtx` should be closed too. That is to say, it's a bug or improvement in pdClient. @leoppro @overvenus