@leoppro 2020-12-17
> > There are two levels of context in PD client: client context and tso request context.
> 
> I think when `clientCtx` is closed, the caller should guarantee the `requestCtx` should be closed too. That is to say, it's not a bug or improvement in pdClient. @leoppro @overvenus

I don't agree with you. Let's think about this following cases:

```
client:=pd.NewClient(xxx) // this line is equivalent to `client:=pd.NewClientWithContext(context.Backend(),xxx)`

// goroutine1
client.GetTSO(context.Backend(), xx)

// goroutine2 
client.Close() // this will lead to client context canceled, and goroutine1 will be blocked
```

this case is exist in tikv client:

here we call GetTSO using backend context: https://github.com/pingcap/tidb/blob/master/store/tikv/kv.go#L387
and here we create pd client using backend context: https://github.com/pingcap/tidb/blob/master/store/tikv/kv.go#L88
note that I can't specify a context for this as a user of tikv client.