@HunDunDM 2020-12-17
I found a similar double `ctx` design in [grpc](https://github.com/grpc/grpc-go/blob/v1.26.0/clientconn.go#L530), and I think it can be merged.